+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/renote/item_renote_user.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.note.view
+package net.pantasystem.milktea.note.renote
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
